Make reticle move with integer precision

+++ gaze_handler.hpp
@@ -30,7 +30,7 @@ class GazeHandler : public QObject, public QRunnable {
 
     signals:
         void socketOut(std::string);
-        void reticleOut(double x, double y);
+        void reticleOut(int x, int y);
         void xmlOut(GazeData gd);
 };
